Add NNAPI support on yocto
This commit adds a layer to enable NNAPI on yocto.

This is adapted from what google did for their chromeos
https://chromium.googlesource.com/aosp/platform/frameworks/ml/+/refs/heads/master/nn/

Makefiles are written by hand and adapted from BUILD.gn. SHA1 of the
different repo are extracted from chromeos build system :
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/d6a2f18398235a000be954954f8c596ad8e4478b/chromeos-base/aosp-frameworks-ml-nn/aosp-frameworks-ml-nn-0.0.1-r166.ebuild

In order to compile, somme patchs need to be applied to add Makefiles
and missing includes

NNAPI should be compiled with clang, so we add the
`TOOLCHAIN = "clang"` directive in the recipe (use meta-clang to get a
clang toolchain)

Signed-off-by: Julien STEPHAN <jstephan@baylibre.com>