nnstreamer: fix test script for object_detection_yolov5
After upgrading to tfv2.14, the converted dla of yolov5s
was different from tfv2.10 and cause an inaccurate
inference result for the `objection_detection_yolov5`.

The converted dla file (from tfv2.10) is directly added to
the demo directory as an workaround to avoid the inaccurate
inference result.

The reason for the conversion difference is to be investigated.

Change-Id: I553e0fd0f0636473621b9c9e535f73149ecee270
Signed-off-by: JoyingKyc <joying.kuo@mediatek.com>